Created metal kickstart and refactoring boots/ hosts

+++ lib/ruby/lib/alces/stack/boot/run.rb
@@ -21,7 +21,7 @@
 #==============================================================================
 require 'alces/tools/logging'
 require 'alces/tools/execution'
-require 'alces/stack/nodes'
+require 'alces/stack/iterator'
 require "alces/stack/templater"
 
 module Alces
@@ -32,53 +32,66 @@ module Alces
         include Alces::Tools::Execution
 
         def initialize(options={})
-          @node_name = options[:name]
-          @group_flag = options[:group_flag]
-          @gender = options[:group]
           @template = options[:template]
+          @group = options[:group]
+          @dry_run_flag = options[:dry_run_flag]
+          @template_parameters = {
+            hostip: `hostname -i`.chomp,
+            nodename: false,
+            kernelappendoptions: options[:kernel_append].chomp
+          }
+          @template_parameters[:nodename] = options[:nodename].chomp if options[:nodename]
+          @json = options[:json]
           @delete_node = ""
-          @kernel_append = options[:kernel_append]
         end
 
         def run!
           puts "(CTRL+C TO TERMINATE)"
+          raise "Requires a node name, node group, or json input" if !@node_name and !@group and !@json 
+
+          case 
+          when @dry_run_flag
+            lambda = -> (json) {puts_template(json)}
+          else
+            lambda = -> (json) {save_template(json)}
+          end
+
           begin
-            if @group_flag
-              run_group
-            else
-              run_single(false)
-            end
+            Alces::Stack::Iterator.new(@group, lambda, @json)
+            sleep
           rescue Exception => e
-            @delete_node.split(',').each do |s|
-              `rm -f /var/lib/tftpboot/pxelinux.cfg/#{s} 2>/dev/null`
+            if @dry_run_flag
+              puts "Would delete the following files:"
+              @delete_node.split(',').each do |s|
+                puts "  /var/lib/tftpboot/pxelinux.cfg/#{s}"
+              end
+            else
+              @delete_node.split(',').each do |s|
+                `rm -f /var/lib/tftpboot/pxelinux.cfg/#{s} 2>/dev/null`
+              end
             end
             raise e
           end
         end
 
-        def run_single(no_hang)
-          raise "No node name supplied" if !@node_name
-          ip=`gethostip -x #{@node_name} 2>/dev/null`
-          raise "Could not find IP address of #{@node_name}" if ip.length < 9
+        def save_template(json)
+          hash = Alces::Stack::Templater::JSON_Templater.parse(json, @template_parameters)
+          ip=`gethostip -x #{hash[:nodename]} 2>/dev/null`
+          raise "Could not find IP address of #{hash[:nodename]}" if ip.length < 9
           @delete_node << ",#{ip}"
-          
           save="/var/lib/tftpboot/pxelinux.cfg/#{ip}"
-          template_parameters = {
-            :hostip => `hostname -i`.chomp,
-            :node => @node_name.chomp,
-            :kernelappendoptions=> @kernel_append.chomp
-          }
-
-          Templater.save(@template, save, template_parameters)
-          sleep if !no_hang
+          Alces::Stack::Templater.save(@template, save, hash)
         end
 
-        def run_group
-          Nodes.new(@gender).each do |node|
-            @node_name = node
-            run_single(true)
-          end
-          sleep
+        def puts_template(json)
+          hash = Alces::Stack::Templater::JSON_Templater.parse(json, @template_parameters)
+          ip=`gethostip -x #{hash[:nodename]} 2>/dev/null`
+          raise "Could not find IP address of #{hash[:nodename]}" if ip.length < 9
+          @delete_node << ",#{ip}"
+          save="/var/lib/tftpboot/pxelinux.cfg/#{ip}"
+          puts "Would save file to: " << save
+          puts Alces::Stack::Templater.file(@template, hash)
+          puts
         end
       end
     end
